pfe_ddos_mitigation/webapp/requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
elasticsearch==9.0.0
paramiko==3.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Bootstrap4==4.0.2
plotly==5.18.0
pandas==2.2.3
